Replaced uint16_t with unsinged

+++ src/cursor.h
@@ -5,8 +5,8 @@
 
 typedef struct Cursor {
 	char canClick;
-	uint16_t yPos, xPos;
-	uint16_t yOld, xOld;
+	unsigned yPos, xPos;
+	unsigned yOld, xOld;
 	char *icon;
 } Cursor;
