+++ src/screen.h
@@ -6,8 +6,8 @@
 #include "cursor.h"
 
 typedef struct Screen {
-	uint16_t depth, maxDepth;  // The amount of layers that there are
-	uint16_t yLength, xLength; // Size of the screen
+	unsigned depth, maxDepth;  // The amount of layers that there are
+	unsigned yLength, xLength; // Size of the screen
 	Layer **layer; // The layers for the screen
 	UpdateStack *update; // The stack for tiles to be drawn
 	Cursor *cursor; // The cursor for interacting with the screen
